+++ extensions/amp-ad-network-adsense-impl/0.1/test/test-amp-ad-network-adsense-impl.js
@@ -21,9 +21,6 @@ import {
 } from '../amp-ad-network-adsense-impl';
 import {AMP_ANALYTICS_HEADER} from '../../../../ads/google/a4a/utils';
 import {
-  setupForAdTesting,
-} from '../../../../extensions/amp-a4a/0.1/test/utils';
-import {
   installExtensionsService,
 } from '../../../../src/service/extensions-impl';
 import {AmpAdUIHandler} from '../../../amp-ad/0.1/amp-ad-ui'; // eslint-disable-line no-unused-vars
@@ -38,6 +35,7 @@ import {
   createElementWithAttributes,
   addAttributesToElement,
 } from '../../../../src/dom';
+import {installDocService} from '../../../../src/service/ampdoc-impl';
 
 function createAdsenseImplElement(attributes, opt_doc, opt_tag) {
   const doc = opt_doc || document;
@@ -46,6 +44,18 @@ function createAdsenseImplElement(attributes, opt_doc, opt_tag) {
     'type': 'adsense',
   });
   return addAttributesToElement(element, attributes);
+}
+
+function setupForAdTesting(fixture) {
+  installDocService(fixture.win, /* isSingleDoc */ true);
+  const doc = fixture.doc;
+  // TODO(a4a-cam@): This is necessary in the short term, until A4A is
+  // smarter about host document styling.  The issue is that it needs to
+  // inherit the AMP runtime style element in order for shadow DOM-enclosed
+  // elements to behave properly.  So we have to set up a minimal one here.
+  const ampStyle = doc.createElement('style');
+  ampStyle.setAttribute('amp-runtime', 'scratch-fortesting');
+  doc.head.appendChild(ampStyle);
 }
 
 describes.sandboxed('amp-ad-network-adsense-impl', {}, () => {
